Clean up and refactor UI and game logic code
After transitioning the game from the console to a GUI, the code had a lot of
functions that didn't have any more use, like player creation factories and
the use of players in functions. As player turns and markers are now decided upon a turn
changer that reacts to user clicks on the gameboard, the game logic didn't have any use
for player creation.
I also cleaned up the GUI for a more presentable look, and while it could
take more work, I undertook this project for excercising Javascript game logic development, more
than anything. So look and feel have been developed to a functional and clean status.